requirements.txt
click==7.1.2
flask-heroku==0.1.9
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
passlib==1.7.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
